# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ovrly API",
    "version": "0.2.0-draft",
    "summary": "Contract between the Android client and the backend.",
    "description": "Ovrly contracts 0.2.0-draft (pre-1.0). Shared-media investigation reads derive state from the intake job (queued/leased -> queued, running -> running, failed -> failed with PROCESSING_FAILED, cancelled/deleted -> cancelled). Capture progress uses GET /v1/captures/{capture_id}; media analysis and claim extraction are not implemented yet. Every schema is a reference into schemas/*.schema.json so the OpenAPI document, the JSON Schema fixtures and the Android models cannot drift from each other. Every error response uses the shared error shape; stage-specific error codes are listed per operation and must be SCREAMING_SNAKE_CASE. Paths listed here are those the backend serves today (BE-05, #19, including account linking from BC-D07, the owner-scoped job actions from BE-04 part 3, #75, and incremental capture intake from BE-06, #24) plus the report versions, explicit saves and POST /v1/voice/actions of BE-10 (#33; the voice contract is BE-13, #67). Reanalysis and the export payload are BE-10 too.",
    "license": {
      "name": "PolyForm Noncommercial 1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development helper (scripts/backend.sh). No hosted server is published by this document."
    }
...
  "security": [{ "bearer": [] }],
...
  "paths": {
...
    "/v1/investigations": {
      "post": {
        "tags": ["investigations"],
        "operationId": "createInvestigation",
        "summary": "Record an investigation durably, then answer 202.",
        "description": "Requires Idempotency-Key. The same key with the same body replays the original 202; the same key with a different body is 409 IDEMPOTENCY_KEY_REUSED. Keys are scoped per owner.",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestigationCreateRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Investigation recorded at stage intake with processing_status waiting.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Investigation" } }
            }
          },
          "400": {
            "description": "IDEMPOTENCY_KEY_REQUIRED or IDEMPOTENCY_KEY_INVALID (longer than 200 characters).",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED (different body) or UPLOAD_INCOMPLETE (the upload is not completed).",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "422": {
            "description": "VALIDATION_FAILED, CLIENT_IDENTITY_REJECTED or DURATION_LIMIT_EXCEEDED.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      },
      "get": {
        "tags": ["investigations"],
        "operationId": "listInvestigations",
        "summary": "List the caller's investigations, newest first, at most 100.",
        "description": "Owner-scoped: only the caller's investigations are visible. Each item is the same read model as GET /v1/investigations/{investigation_id}.",
        "responses": {
          "200": {
            "description": "Owner-scoped list.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvestigationList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      }
    },
...
    "/v1/investigations/{investigation_id}": {
      "parameters": [{ "$ref": "#/components/parameters/InvestigationId" }],
      "get": {
        "tags": ["investigations"],
        "operationId": "getInvestigation",
        "summary": "Read one investigation.",
        "description": "Another owner's investigation is a plain 404; existence is not revealed. The six result fixtures under fixtures/results are payloads of this response.",
        "responses": {
          "200": {
            "description": "Investigation read model.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Investigation" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationFailed" },
          "500": { "$ref": "#/components/responses/InternalError" },
          "503": { "$ref": "#/components/responses/DatabaseUnavailable" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque token minted by POST /v1/principals/guest (ovk_ prefix). Identity comes only from this credential; a user_id, owner_id or principal_id in a body, query or header is rejected with CLIENT_IDENTITY_REJECTED."
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Echoed from a well-formed client X-Request-Id (opaque_id syntax) or generated. Equals error.request_id on every error response.",
        "schema": { "$ref": "schemas/common.schema.json#/$defs/opaque_id" }
      }
    },
    "parameters": {
...
      "InvestigationId": {
        "name": "investigation_id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "schemas/common.schema.json#/$defs/uuid" }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Client-chosen key, at most 200 characters after trimming, scoped to the caller.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
      }
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "AUTHENTICATION_REQUIRED (no credential) or INVALID_CREDENTIAL (malformed, unknown or revoked), with WWW-Authenticate.",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "WWW-Authenticate": { "schema": { "type": "string" } }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "NOT_FOUND: no such object visible to the caller. Another owner's object answers the same way.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ValidationFailed": {
        "description": "VALIDATION_FAILED (body, query or path invalid) or CLIENT_IDENTITY_REJECTED.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "InternalError": {
        "description": "INTERNAL_ERROR: unexpected failure; no internals are revealed.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "DatabaseUnavailable": {
        "description": "DATABASE_UNAVAILABLE: retryable; action retry.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
...
      "InvestigationList": {
        "type": "object",
        "description": "Mirrors services/api/schemas.py InvestigationListResponse.",
        "additionalProperties": false,
        "required": ["items"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Investigation" } }
        }
      },
      "Error": { "$ref": "schemas/error.schema.json" },
...
      "InvestigationCreateRequest": { "$ref": "schemas/investigation-create-request.schema.json" },
      "Investigation": { "$ref": "schemas/investigation.schema.json" },
...
    }
  }
}